requirements.txt
# ============================================================
# Telegram API Checker Bot - Dependencies
# ============================================================
# Version: 1.0.0
# Python: >=3.8
# ============================================================

# Core Telegram Bot Framework
python-telegram-bot[all]>=20.7
python-telegram-bot-raw>=20.7

# Web Automation & Browser Control
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.1
pyppeteer>=1.0.2

# HTML/XML Parsing & Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
parsel>=1.8.1

# HTTP Clients & Networking
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0
urllib3>=2.1.0
requests-html>=0.10.0

# Async Support
asyncio>=3.4.3
nest-asyncio>=1.5.8
aiofiles>=23.2.1
aiodns>=3.1.1

# Data Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
marshmallow>=3.20.1

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0.1
configparser>=6.0.0
python-decouple>=3.8

# Database & ORM
sqlalchemy>=2.0.23
alembic>=1.13.0
aiosqlite>=0.19.0
databases>=0.8.0

# Template Engine
jinja2>=3.1.2
mako>=1.3.0

# Security & Cryptography
cryptography>=41.0.7
pycryptodome>=3.19.0
bcrypt>=4.1.1
passlib>=1.7.4

# Image Processing
pillow>=10.1.0
opencv-python>=4.8.1.78

# CLI & Terminal UI
colorama>=0.4.6
rich>=13.7.0
tqdm>=4.66.1
click>=8.1.7
typer>=0.9.0

# Logging & Monitoring
loguru>=0.7.2
python-json-logger>=2.0.7

# User Agent & Headers
fake-useragent>=1.4.0
user-agents>=2.2.0

# JSON & Data Processing
jsonschema>=4.20.0
ujson>=5.9.0
orjson>=3.9.10
python-rapidjson>=1.13

# Regular Expressions & Pattern Matching
regex>=2023.10.3

# Rate Limiting & Throttling
ratelimit>=2.2.1
aiolimiter>=1.1.0

# Proxy Support
pysocks>=1.7.1
requests[socks]>=2.31.0

# JWT & Token Processing
pyjwt>=2.8.0
python-jose>=3.3.0

# GraphQL Support
gql>=3.5.0
graphql-core>=3.2.3

# API Testing & Fuzzing
hypothesis>=6.92.0

# Date & Time Processing
python-dateutil>=2.8.2
pytz>=2023.3
arrow>=1.3.0

# File Operations & Compression
zipfile38>=0.0.3
py7zr>=0.20.8

# Code Generation
black>=23.12.0
autopep8>=2.0.4

# Testing & Quality Assurance
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0

# Type Checking
mypy>=1.7.0
types-requests>=2.31.0
types-pyyaml>=6.0.12

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0

# Development Tools
ipython>=8.18.0
ipdb>=0.13.13

# Environment Detection
python-magic>=0.4.27

# Misc Utilities
tenacity>=8.2.3
backoff>=2.2.1
retrying>=1.3.4
